Extra search filter logic to service
